+++ packages/react/src/deprecated/index.ts
@@ -43,7 +43,6 @@ export {default as Tooltip} from '../Tooltip/Tooltip'
 export type {TooltipProps} from '../Tooltip/Tooltip'
 // end of v37.0.0
 
-// Deprecated in v37.31.0 - https://github.com/github/primer/issues/5533
-export {default as AvatarPair} from './AvatarPair'
-export type {AvatarPairProps} from './AvatarPair'
-// end of v37.31.0
+// Deprecated in v38.0.0
+export {default as PointerBox} from '../PointerBox'
+export type {PointerBoxProps} from '../PointerBox'