+++ packages/styled-react/src/experimental.ts
@@ -1,1 +1,0 @@
-export {Dialog} from '@primer/react/experimental'