+++ packages/styled-react/src/experimental.tsx
@@ -1,0 +1,1 @@
+export {Dialog, PageHeader, Table, Tooltip, UnderlinePanels} from '@primer/react/experimental'